Add dispatcher diversion to resolve
