--- internal/server/constants.go
@@ -1,41 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package server
-
-const (
-	// --- Constants related to the resolve endpoint ---
-	// ResolveTargetBaseOnly indicates resolution should only use the base mixer.
-	ResolveTargetBaseOnly = "base_only"
-	// ResolveTargetCustomOnly indicates resolution should only use the custom mixer.
-	ResolveTargetCustomOnly = "custom_only"
-	// ResolveTargetBaseAndCustom indicates resolution should use both custom and base mixers.
-	ResolveTargetBaseAndCustom = "base_and_custom"
-
-	// ResolveResolverPlace is the default resolver type.
-	ResolveResolverPlace = "place"
-	// ResolveResolverIndicator is the resolver name for indicator/embeddings resolution.
-	ResolveResolverIndicator = "indicator"
-
-	// ResolveDefaultPropertyExpression is the property name for description.
-	ResolveDefaultPropertyExpression = "<-description->dcid"
-	// DcidProperty is the property name for dcid.
-	DcidProperty = "dcid"
-	// DescriptionProperty is the property name for description.
-	DescriptionProperty = "description"
-	// GeoCoordinateProperty is the property name for geoCoordinate.
-	GeoCoordinateProperty = "geoCoordinate"
-	// TypeOfProperty is the property name for typeOf.
-	TypeOfProperty = "typeOf"
-)
