+++ internal/server/v2/resolve/resolve.go
@@ -17,6 +17,7 @@ package resolve
 
 import (
 	"context"
+	"fmt"
 	"sort"
 	"strconv"
 	"strings"
@@ -25,9 +26,36 @@ import (
 	pb "github.com/datacommonsorg/mixer/internal/proto"
 	pbv2 "github.com/datacommonsorg/mixer/internal/proto/v2"
 	"github.com/datacommonsorg/mixer/internal/server/recon"
+	v2 "github.com/datacommonsorg/mixer/internal/server/v2"
 	"github.com/datacommonsorg/mixer/internal/store"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+)
+
+const (
+	// --- Constants related to the resolve endpoint ---
+	// ResolveTargetBaseOnly indicates resolution should only use the base mixer.
+	ResolveTargetBaseOnly = "base_only"
+	// ResolveTargetCustomOnly indicates resolution should only use the custom mixer.
+	ResolveTargetCustomOnly = "custom_only"
+	// ResolveTargetBaseAndCustom indicates resolution should use both custom and base mixers.
+	ResolveTargetBaseAndCustom = "base_and_custom"
+
+	// ResolveResolverPlace is the default resolver type.
+	ResolveResolverPlace = "place"
+	// ResolveResolverIndicator is the resolver name for indicator/embeddings resolution.
+	ResolveResolverIndicator = "indicator"
+
+	// ResolveDefaultPropertyExpression is the property name for description.
+	ResolveDefaultPropertyExpression = "<-description->dcid"
+	// DcidProperty is the property name for dcid.
+	DcidProperty = "dcid"
+	// DescriptionProperty is the property name for description.
+	DescriptionProperty = "description"
+	// GeoCoordinateProperty is the property name for geoCoordinate.
+	GeoCoordinateProperty = "geoCoordinate"
+	// TypeOfProperty is the property name for typeOf.
+	TypeOfProperty = "typeOf"
 )
 
 var resolvedPlaceTypePriorityList = []string{
@@ -236,4 +264,169 @@ func GetSortedResolvedPlaceCandidates(
 	candidates = append(candidates, leftoverCandidates...)
 
 	return candidates
+}
+
+// validateAndParseResolveInputs validates and parses the inputs for the resolve request.
+//
+// Validation logic:
+// - `target`: Must be one of "base_only", "custom_only", "base_and_custom". Defaults to "base_and_custom".
+// - `resolver`: Must be one of "place", "indicator". Defaults to "place".
+// - `property`:
+//   - Must match the format "<-inProp->outProp" (optionally with filters).
+//   - "inProp" and "outProp" validation depends on the resolver:
+//   - "place": if "inProp" is "description" or "geoCoordinate", "outProp" must be "dcid".
+//   - "indicator": "inProp" must be "description" and "outProp" must be "dcid".
+//
+// Returns:
+// - input property (string)
+// - output property (string)
+// - typeOf filter values ([]string, from the input property filter)
+// - error if validation fails
+func ValidateAndParseResolveInputs(in *pbv2.ResolveRequest) (string, string, []string, error) {
+	var validationErrors []string
+
+	// Parse and validate `target`
+	switch in.GetTarget() {
+	case ResolveTargetBaseOnly, ResolveTargetCustomOnly, ResolveTargetBaseAndCustom:
+	case "":
+		// Set default value; ignored if current call is to base dc
+		in.Target = ResolveTargetBaseAndCustom
+	default:
+		validationErrors = append(validationErrors, fmt.Sprintf("Invalid 'target': valid values are '%s', '%s', '%s'",
+			ResolveTargetCustomOnly, ResolveTargetBaseOnly, ResolveTargetBaseAndCustom))
+	}
+
+	// Parse and validate `resolver`
+	switch in.GetResolver() {
+	case ResolveResolverPlace, ResolveResolverIndicator:
+	case "":
+		// Set default value
+		in.Resolver = ResolveResolverPlace
+	default:
+		validationErrors = append(validationErrors, fmt.Sprintf("Invalid 'resolver': valid values are '%s', '%s'",
+			ResolveResolverIndicator, ResolveResolverPlace))
+	}
+
+	// Parse `property` expression into its in arc property, out arc property and typeOf filter values
+	if in.GetProperty() == "" {
+		in.Property = ResolveDefaultPropertyExpression
+	}
+
+	inProp, outProp, typeOfValues, err := parseResolvePropertyExpression(in.GetProperty())
+	if err != nil {
+		validationErrors = append(validationErrors, fmt.Sprintf("Invalid 'property' expression: %v", err))
+	}
+
+	// Validate property expression based on resolver
+	if err == nil {
+		switch in.GetResolver() {
+		case ResolveResolverPlace:
+			// geoCoordinate and description only support dcid as outArc
+			if (inProp == DescriptionProperty || inProp == GeoCoordinateProperty) && outProp != DcidProperty {
+				validationErrors = append(validationErrors, fmt.Sprintf(
+					"Invalid 'property' expression: given input property '%s', output property can only be '%s'",
+					inProp, DcidProperty))
+			}
+		case ResolveResolverIndicator:
+			// Indicator resolution only supports description as inArc.
+			if inProp != DescriptionProperty {
+				validationErrors = append(validationErrors, fmt.Sprintf(
+					"Invalid 'property' expression: indicator resolution only supports '%s' as input property",
+					DescriptionProperty))
+			}
+			// Indicator resolution only supports dcid as outArc.
+			if outProp != DcidProperty {
+				validationErrors = append(validationErrors, fmt.Sprintf(
+					"Invalid 'property' expression: indicator resolution only supports '%s' as output property",
+					DcidProperty))
+			}
+		}
+	}
+
+	if len(validationErrors) > 0 {
+		return "", "", nil, status.Errorf(codes.InvalidArgument, "Invalid inputs in request. %s", strings.Join(validationErrors, ". "))
+	}
+
+	return inProp, outProp, typeOfValues, nil
+}
+
+// parseResolvePropertyExpression parses and validates a property expression string.
+//
+// The expression generally takes the form "<-inProp->outProp", optionally with filters
+// like "<-inProp{typeOf:Type}->outProp".
+//
+// Returns:
+// - input property (string)
+// - output property (string)
+// - typeOf filter values ([]string, from the input property filter)
+// - error if validation fails
+func parseResolvePropertyExpression(prop string) (string, string, []string, error) {
+	// Parse property expression into Arcs.
+	arcs, err := v2.ParseProperty(prop)
+	if err != nil {
+		return "", "", nil, err
+	}
+
+	if len(arcs) != 2 {
+		return "", "", nil, fmt.Errorf("must define exactly two parts (incoming and outgoing arcs). Found %d parts", len(arcs))
+	}
+
+	inArc := arcs[0]
+	outArc := arcs[1]
+	if inArc.Out || !outArc.Out {
+		return "", "", nil, fmt.Errorf("must start with an incoming arc and end with an outgoing arc")
+	}
+
+	if inArc.SingleProp == "" {
+		return "", "", nil, fmt.Errorf("input property must be provided")
+	}
+	if outArc.SingleProp == "" {
+		return "", "", nil, fmt.Errorf("output property must be provided")
+	}
+
+	var typeOfValues []string
+	// Validate filters
+	if len(inArc.Filter) > 0 {
+		if len(inArc.Filter) > 1 {
+			return "", "", nil, fmt.Errorf("only '%s' filter is supported", TypeOfProperty)
+		}
+		if filter, ok := inArc.Filter[TypeOfProperty]; !ok {
+			for k := range inArc.Filter {
+				return "", "", nil, fmt.Errorf("invalid filter key '%s'. Only '%s' filter is supported", k, TypeOfProperty)
+			}
+		} else {
+			typeOfValues = filter
+		}
+	}
+
+	return inArc.SingleProp, outArc.SingleProp, typeOfValues, nil
+}
+
+// resolveRouting determines whether to route to local and/or remote instances
+// based on the target parameter and the presence of a remote mixer domain.
+// Returns (shouldCallLocal, shouldCallRemote, error).
+//
+// Assumes that `target` has been validated.
+//
+// logic:
+//   - If remoteMixerDomain is empty, we are the base instance (or standalone).
+//     Always process locally, ignore target.
+//   - If remoteMixerDomain is set, we are a custom instance.
+//     Route based on target:
+//   - "base_only": Call remote only.
+//   - "custom_only": Call local only.
+//   - "base_and_custom": Call both.
+//   - Any other value defaults to calling both.
+func ResolveRouting(target string, hasRemoteMixerDomain bool) (bool, bool) {
+	if !hasRemoteMixerDomain {
+		return true, false
+	}
+	switch target {
+	case ResolveTargetBaseOnly:
+		return false, true
+	case ResolveTargetCustomOnly:
+		return true, false
+	default:
+		return true, true
+	}
 }
