+++ internal/server/v2/resolve/resolve_test.go
@@ -15,7 +15,13 @@
 // Package resolve is for V2 resolve API.
 package resolve
 
-import "testing"
+import (
+	"strings"
+	"testing"
+
+	pbv2 "github.com/datacommonsorg/mixer/internal/proto/v2"
+	"github.com/google/go-cmp/cmp"
+)
 
 func TestParseCoordinate(t *testing.T) {
 	for _, c := range []struct {
@@ -44,5 +50,220 @@ func TestParseCoordinate(t *testing.T) {
 			t.Errorf("parseCoordinate(%s) = %f, %f, want %f, %f",
 				c.coordinateExpr, gotLat, gotLng, c.wantLat, c.wantLng)
 		}
+	}
+}
+
+func TestResolveRouting(t *testing.T) {
+	tests := []struct {
+		desc                 string
+		target               string
+		hasRemoteMixerDomain bool
+		wantLocal            bool
+		wantRemote           bool
+		wantErr              bool
+	}{
+		{
+			desc:       "Base instance (empty remote domain)",
+			target:     "any_target",
+			wantLocal:  true,
+			wantRemote: false,
+			wantErr:    false,
+		},
+		{
+			desc:                 "Custom instance, target base_only",
+			target:               ResolveTargetBaseOnly,
+			hasRemoteMixerDomain: true,
+			wantLocal:            false,
+			wantRemote:           true,
+			wantErr:              false,
+		},
+		{
+			desc:                 "Custom instance, target custom_only",
+			target:               ResolveTargetCustomOnly,
+			hasRemoteMixerDomain: true,
+			wantLocal:            true,
+			wantRemote:           false,
+			wantErr:              false,
+		},
+		{
+			desc:                 "Custom instance, target base_and_custom",
+			target:               ResolveTargetBaseAndCustom,
+			hasRemoteMixerDomain: true,
+			wantLocal:            true,
+			wantRemote:           true,
+			wantErr:              false,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.desc, func(t *testing.T) {
+			gotLocal, gotRemote := ResolveRouting(tc.target, tc.hasRemoteMixerDomain)
+			if !tc.wantErr {
+				if gotLocal != tc.wantLocal || gotRemote != tc.wantRemote {
+					t.Errorf("resolveRouting(%q, %t) = (%v, %v), want (%v, %v)",
+						tc.target, tc.hasRemoteMixerDomain, gotLocal, gotRemote, tc.wantLocal, tc.wantRemote)
+				}
+			}
+		})
+	}
+}
+
+func TestValidateAndParseResolveInputs(t *testing.T) {
+	tests := []struct {
+		desc             string
+		in               *pbv2.ResolveRequest
+		wantInProp       string
+		wantOutProp      string
+		wantTypeOfValues []string
+		wantErr          bool
+		wantErrMsg       string
+	}{
+		{
+			desc:             "all empty",
+			in:               &pbv2.ResolveRequest{},
+			wantInProp:       "description",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: nil,
+		},
+		{
+			desc: "partial set - target",
+			in: &pbv2.ResolveRequest{
+				Target: ResolveTargetCustomOnly,
+			},
+			wantInProp:       "description",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: nil,
+		},
+		{
+			desc: "partial set - resolver",
+			in: &pbv2.ResolveRequest{
+				Resolver: ResolveResolverIndicator,
+			},
+			wantInProp:       "description",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: nil,
+		},
+		{
+			desc: "fully set",
+			in: &pbv2.ResolveRequest{
+				Target:   ResolveTargetBaseOnly,
+				Resolver: ResolveResolverPlace,
+				Property: "<-wikidataId->dcid",
+			},
+			wantInProp:       "wikidataId",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: nil,
+		},
+		{
+			desc: "typeOf filter",
+			in: &pbv2.ResolveRequest{
+				Property: "<-description{typeOf:City}->dcid",
+			},
+			wantInProp:       "description",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: []string{"City"},
+		},
+		{
+			desc: "invalid target",
+			in: &pbv2.ResolveRequest{
+				Target: "invalid",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'target': valid values are 'custom_only', 'base_only', 'base_and_custom'",
+		},
+		{
+			desc: "invalid resolver",
+			in: &pbv2.ResolveRequest{
+				Resolver: "invalid",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'resolver': valid values are 'indicator', 'place'",
+		},
+		{
+			desc: "invalid target and resolver",
+			in: &pbv2.ResolveRequest{
+				Target:   "invalid_target",
+				Resolver: "invalid_resolver",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'target': valid values are 'custom_only', 'base_only', 'base_and_custom'. Invalid 'resolver': valid values are 'indicator', 'place'",
+		},
+		{
+			desc: "invalid property expression",
+			in: &pbv2.ResolveRequest{
+				Property: "invalid_prop",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'property' expression",
+		},
+		{
+			desc: "unknown property for place resolver",
+			in: &pbv2.ResolveRequest{
+				Resolver: ResolveResolverPlace,
+				Property: "<-unknown->dcid",
+			},
+			wantInProp:       "unknown",
+			wantOutProp:      "dcid",
+			wantTypeOfValues: nil,
+		},
+		{
+			desc: "invalid property for indicator resolver (inProp)",
+			in: &pbv2.ResolveRequest{
+				Resolver: ResolveResolverIndicator,
+				Property: "<-geoCoordinate->dcid",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'property' expression: indicator resolution only supports 'description' as input property",
+		},
+		{
+			desc: "invalid property for indicator resolver (outProp)",
+			in: &pbv2.ResolveRequest{
+				Resolver: ResolveResolverIndicator,
+				Property: "<-description->nutsCode",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'property' expression: indicator resolution only supports 'dcid' as output property",
+		},
+		{
+			desc: "invalid target + valid unknown property",
+			in: &pbv2.ResolveRequest{
+				Target:   "invalid",
+				Resolver: ResolveResolverPlace,
+				Property: "<-unknown->dcid",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'target': valid values are 'custom_only', 'base_only', 'base_and_custom'",
+		},
+		{
+			desc: "invalid output property for place (description)",
+			in: &pbv2.ResolveRequest{
+				Property: "<-description->nutsCode",
+			},
+			wantErr:    true,
+			wantErrMsg: "Invalid inputs in request. Invalid 'property' expression: given input property 'description', output property can only be 'dcid'",
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.desc, func(t *testing.T) {
+			inProp, outProp, typeOfValues, err := ValidateAndParseResolveInputs(tc.in)
+			if (err != nil) != tc.wantErr {
+				t.Errorf("validateAndParseResolveInputs() error = %v, wantErr %v", err, tc.wantErr)
+			}
+			if tc.wantErr && err != nil && !strings.Contains(err.Error(), tc.wantErrMsg) {
+				t.Errorf("validateAndParseResolveInputs() error = %v, wantErrMsg %v", err, tc.wantErrMsg)
+			}
+			if !tc.wantErr {
+				if inProp != tc.wantInProp {
+					t.Errorf("inProp got %s, want %s", inProp, tc.wantInProp)
+				}
+				if outProp != tc.wantOutProp {
+					t.Errorf("outProp got %s, want %s", outProp, tc.wantOutProp)
+				}
+				if diff := cmp.Diff(typeOfValues, tc.wantTypeOfValues); diff != "" {
+					t.Errorf("typeOfValues diff (-got +want):\n%s", diff)
+				}
+			}
+		})
 	}
 }
