+++ internal/server/v3/resolve/golden/resolve_description_type_filter.json
@@ -1,7 +1,12 @@
 {
   "entities": [
     {
-      "node": "California"
+      "node": "California",
+      "candidates": [
+        {
+          "dcid": "geoId/06"
+        }
+      ]
     }
   ]
 }
